+++ CHANGES.md
@@ -16,6 +16,7 @@ Changes in 4.3
 - CSS Add .wkhtmltopdf-header, .wkhtmltopdf-footer, .wkhtmltopdf-portrait & .wkhtmltopdf-landscape classes in PDF.php & wkhtmltopdf.css
 - Fix SQL error for FOOD_SERVICE_ACCOUNT table when adding student reusing deleted student ID in Student.php
 - Fix SQL error when Parent user exists and is already associated to student in CreateParents.php
+- Add SubstitutionsInput & SubstitutionsTextMake functions in Substitutions.fnc.php
 
 Changes in 4.2
 --------------